Refactor e2e
